pyproject.toml
[project]
name = "practica-servicios"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-phonenumber-field[phonenumbers]>=8.4.0",
]
